+++ market-manager-service/csb-service-api/src/main/java/com/arman/csb/modules/service/ScoreLocalService.java
@@ -288,7 +288,7 @@ public interface ScoreLocalService extends BaseLocalService,
         java.lang.String[] parameterTypes, java.lang.Object[] arguments)
         throws java.lang.Throwable;
 
-    public int sumByCustomerAndType(java.lang.Long customerId,
+    public long sumByCustomerAndType(java.lang.Long customerId,
         java.lang.Integer type, java.util.Date fromDate, java.util.Date toDate)
         throws com.liferay.portal.kernel.exception.PortalException,
             com.liferay.portal.kernel.exception.SystemException;
@@ -301,7 +301,7 @@ public interface ScoreLocalService extends BaseLocalService,
     * @throws SystemException
     * @throws PortalException
     */
-    public int sumByType(java.lang.Integer type, java.util.Date fromDate,
+    public long sumByType(java.lang.Integer type, java.util.Date fromDate,
         java.util.Date toDate)
         throws com.liferay.portal.kernel.exception.PortalException,
             com.liferay.portal.kernel.exception.SystemException;
